+++ MMOTFG_Bot/MMOTFG_Bot/Battle/Level/LevelUpRoadmap.cs
@@ -13,7 +13,7 @@ namespace MMOTFG_Bot
     {
         //base experience needed to level up
         [DefaultValue(1)]
-        public float levelUpExperience;
+        public int levelUpExperience;
 
         //1 for linear, 2 for cuadratic, 3 for cubic...
         [DefaultValue(1)]
@@ -25,9 +25,9 @@ namespace MMOTFG_Bot
         public int maxLevel;
 
         //stats at min level
-        public float[] firstStats;
+        public int[] firstStats;
         //stats at max level
-        public float[] finalStats;
+        public int[] finalStats;
 
         public struct LevelUpEvent
         {
